Flatten config, merge it finally with CLI one
