+++ local-cli/core/android/findManifest.js
