+++ local-cli/core/android/index.js
