+++ local-cli/core/android/readManifest.js
