+++ local-cli/core/findAssets.js
