+++ local-cli/core/ios/index.js
