+++ local-cli/core/windows/findNamespace.js
