+++ local-cli/core/windows/findPackageClassName.js
