+++ local-cli/core/windows/findProject.js
