+++ local-cli/core/windows/generateGUID.js
